pyproject.toml
[project]
name = "floorplan-set-transformer"
version = "0.1.0"
description = "Set Transformer for Panoptic Symbol Spotting on FloorPlanCAD"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "svgpathtools>=1.6.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
    "gdown>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
